修改数据库表 生成生成xml comment/oderDetails表字段更新

+++ src/test/java/com/team6/dao/UserMapperTest.java
@@ -17,7 +17,7 @@ public class UserMapperTest {
     private UserMapper userMapper;
     @Test
     public void queryById() {
-        long id = 10001;
+        int id = 10001;
         User user = userMapper.queryById(id);
         System.out.println(user);
     }
